+++ example/fmm-coulomb/src/utils.cpp
@@ -1,12 +1,13 @@
-unsigned int TriangleNumbers(unsigned int n) {
+#include "utils.hpp"
+
+size_t TriangleNumbers(size_t n) {
 	return (n * (n + 1)) / 2;
 }
 
-unsigned int Nterms(unsigned int p) {
-	unsigned int result = 0;
-	for(unsigned int i = 0; i < p + 2; i++) {
+size_t Nterms(size_t p) {
+	size_t result = 0;
+	for(size_t i = 0; i < p + 2; i++) {
 		result += TriangleNumbers(i);
 	}
 	return result;
 }
-
